requirements.txt
pdfplumber
python-pptx
sentence-transformers
pillow
genanki
spacy
re